apitracking/x/apitracking/types/messages_apirequest_test.go
package types

import (
	"testing"

	"apitracking/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateApirequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateApirequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateApirequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateApirequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateApirequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateApirequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateApirequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateApirequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteApirequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteApirequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteApirequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteApirequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
